Add typings for bumpSequence and move to BaseOperation.
